+++ HISTORY.rst
@@ -121,3 +121,23 @@ History
 0.2.8 (2022-6-20)
 -------------------
 * [update] modify exception handle
+
+0.2.9 (2022-7-20)
+-------------------
+* [update] modify sentence preprocess
+
+0.2.10 (2022-7-27)
+-------------------
+* [update] update classifier
+
+0.2.11 (2022-7-27)
+-------------------
+* [update] bug fix
+
+0.2.12 (2022-7-27)
+-------------------
+* [update] add clean value
+
+0.2.13 (2022-7-31)
+-------------------
+* [update] add concept extractor